--- a/package.json
+++ b/package.json
@@ -20,7 +20,7 @@
   "author": "Chakrit Wichian <service@chakrit.net> (http://chakrit.net)",
   "license": "BSD",
   "dependencies": {
-    "readable-stream": "~1.1.0",
+    "readable-stream": "~1.0.x",
     "throttle": "~1.0.3"
   }
 }